+++ src/JestProcessManagement/JestProcess.ts
@@ -1,19 +1,18 @@
 import { platform } from 'os'
 import { Runner, ProjectWorkspace } from 'jest-editor-support'
 import { WatchMode } from '../Jest'
-
-export type ExitCallback = (process: JestProcess) => void
+import { ExitCallback } from './JestProcessManager'
 
 export class JestProcess {
   static readonly keepAliveLimit = 5
+  static readonly stopHangTimeout = 500
   public keepAlive: boolean
-  public stopRequested: boolean
   public watchMode: WatchMode
   private runner: Runner
   private projectWorkspace: ProjectWorkspace
   private onExitCallback: ExitCallback
   private jestSupportEvents: Map<string, (...args: any[]) => void>
-  private resolve: () => void
+  private stopResolveCallback: () => void | null
   private keepAliveCounter: number
 
   constructor({
@@ -38,28 +37,35 @@ export class JestProcess {
     this.onExitCallback = callback
   }
 
-  public onJestEditorSupportEvent(event, callback) {
+  public onJestEditorSupportEvent(event: string, callback: (...args: any[]) => void) {
     this.jestSupportEvents.set(event, callback)
     this.runner.on(event, callback)
     return this
   }
 
-  public stop() {
-    this.stopRequested = true
-    this.keepAliveCounter = 1
-    this.jestSupportEvents.clear()
-    this.runner.closeProcess()
+  public stop(): Promise<void> {
     return new Promise(resolve => {
-      this.resolve = resolve
+      this.keepAliveCounter = 1
+      this.stopResolveCallback = resolve
+      this.jestSupportEvents.clear()
+      this.runner.closeProcess()
+
+      // As a safety fallback to prevent the stop from hanging, resolve after a timeout
+      // this is safe since subsequent resolve calls are no-op
+      // TODO: If `closeProcess` can be guarenteed to always resolve, remove this
+      setTimeout(resolve, JestProcess.stopHangTimeout)
     })
   }
 
-  public runJestWithUpdateForSnapshots(callback) {
+  public runJestWithUpdateForSnapshots(callback: () => void) {
     this.runner.runJestWithUpdateForSnapshots(callback)
   }
 
+  public stopRequested(): boolean {
+    return this.stopResolveCallback !== null
+  }
   private startRunner() {
-    this.stopRequested = false
+    this.stopResolveCallback = null
     let exited = false
 
     const options = {
@@ -78,10 +84,13 @@ export class JestProcess {
         if (--this.keepAliveCounter > 0) {
           this.runner.removeAllListeners()
           this.startRunner()
-        } else if (this.onExitCallback) {
-          this.onExitCallback(this)
-          if (this.stopRequested) {
-            this.resolve()
+        } else {
+          if (this.onExitCallback) {
+            this.onExitCallback(this)
+          }
+          if (this.stopResolveCallback) {
+            this.stopResolveCallback()
+            this.stopResolveCallback = null
           }
         }
       }